Removed a few linefeeds from the params.pp
While puppet-lint warns about lines longer than 80 characters, trying to avoid
them with forced linefeeds ("\") will actually cause problems in certain cases.
For example, if a string variable is split to several lines and then passed on
to an ERB template, the string in the template will have useless and potentially
harmful extra spaces in it. For this reason this commit reverts some of the
forced linefeeds added in the previous commit.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>